Bottom Tabs and upgrade react navigation
