+++ package.json
@@ -18,7 +18,7 @@
     "@react-navigation/drawer": "^6.1.4",
     "@react-navigation/material-bottom-tabs": "^6.0.9",
     "@react-navigation/native": "^6.0.6",
-    "@react-navigation/native-stack": "^6.1.0",
+    "@react-navigation/native-stack": "^6.2.5",
     "@react-navigation/stack": "^6.0.7",
     "expo-constants": "^12.1.3",
     "expo-font": "^10.0.3",
